+++ src/unstract/sdk/adapters/__init__.py
@@ -1,5 +1,3 @@
-__version__ = "0.20.1"
-
 import logging
 from logging import NullHandler
 from typing import Any
@@ -8,7 +6,3 @@ logging.getLogger(__name__).addHandler(NullHandler())
 
 AdapterDict = dict[str, dict[str, Any]]
 
-
-def get_adapter_version() -> str:
-    """Returns the adapter package's version."""
-    return __version__
