+++ src/unstract/sdk/llm.py
@@ -11,11 +11,11 @@ from unstract.sdk.adapters.constants import Common
 from unstract.sdk.adapters.llm import adapters
 from unstract.sdk.adapters.llm.llm_adapter import LLMAdapter
 
-from unstract.sdk.core.adapters import ToolAdapter
-from unstract.sdk.core.constants import LogLevel, ToolEnv
-from unstract.sdk.core.exceptions import LLMError, RateLimitError, SdkError
-from unstract.sdk.core.tool.base import BaseTool
-from unstract.sdk.core.utils.callback_manager import CallbackManager
+from unstract.sdk.adapters import ToolAdapter
+from unstract.sdk.constants import LogLevel, ToolEnv
+from unstract.sdk.exceptions import LLMError, RateLimitError, SdkError
+from unstract.sdk.tool.base import BaseTool
+from unstract.sdk.utils.callback_manager import CallbackManager
 
 logger = logging.getLogger(__name__)
