+++ src/unstract/sdk/platform.py
@@ -2,9 +2,9 @@ from typing import Any, Optional
 
 import requests
 
-from unstract.sdk.core.constants import LogLevel, ToolEnv
-from unstract.sdk.core.helper import SdkHelper
-from unstract.sdk.core.tool.base import BaseTool
+from unstract.sdk.constants import LogLevel, ToolEnv
+from unstract.sdk.helper import SdkHelper
+from unstract.sdk.tool.base import BaseTool
 
 
 class PlatformBase:
