+++ src/unstract/sdk/prompt.py
@@ -4,9 +4,9 @@ from typing import Any, Optional
 import requests
 from requests import ConnectionError, RequestException, Response
 
-from unstract.sdk.core.constants import LogLevel, PromptStudioKeys, ToolEnv
-from unstract.sdk.core.helper import SdkHelper
-from unstract.sdk.core.tool.base import BaseTool
+from unstract.sdk.constants import LogLevel, PromptStudioKeys, ToolEnv
+from unstract.sdk.helper import SdkHelper
+from unstract.sdk.tool.base import BaseTool
 
 logger = logging.getLogger(__name__)
