+++ src/unstract/sdk/static/tool_template/v1/Dockerfile
