+++ src/unstract/sdk/static/tool_template/v1/data_dir/INFILE
