+++ src/unstract/sdk/static/tool_template/v1/src/config/spec.json
