+++ src/unstract/sdk/tool/stream.py
@@ -5,7 +5,7 @@ from typing import Any
 
 from deprecated import deprecated
 
-from unstract.sdk.core.constants import Command, LogLevel, LogStage, ToolEnv
+from unstract.sdk.constants import Command, LogLevel, LogStage, ToolEnv
 
 
 class StreamMixin:
