+++ src/unstract/sdk/utils/usage_handler.py
@@ -5,10 +5,10 @@ from llama_index.core.callbacks.base_handler import BaseCallbackHandler
 from llama_index.core.embeddings import BaseEmbedding
 from llama_index.core.llms import LLM
 
-from unstract.sdk.core.audit import Audit
-from unstract.sdk.core.constants import LogLevel
-from unstract.sdk.core.tool.stream import StreamMixin
-from unstract.sdk.core.utils.token_counter import TokenCounter
+from unstract.sdk.audit import Audit
+from unstract.sdk.constants import LogLevel
+from unstract.sdk.tool.stream import StreamMixin
+from unstract.sdk.utils.token_counter import TokenCounter
 
 
 class UsageHandler(StreamMixin, BaseCallbackHandler):
